握手收到回复数据成功（之前以为未收到数据，实际上市收到的数据前4个字节可能为0x00，转成字符串时被当做字符串的终止符）

+++ test_cocos_3/MyCppGame/Classes/ODSocket/ODSocket.cpp
@@ -137,7 +137,7 @@ int ODSocket::Send(const char* buf, int len, int flags)
 
 	while ( count < len ) {
 
-		bytes = send(m_sock, buf + count, len - count, flags);
+		bytes = (int)send(m_sock, buf + count, len - count, flags);
 		if ( bytes == -1 || bytes == 0 )
 			return -1;
 		count += bytes;
@@ -148,7 +148,7 @@ int ODSocket::Send(const char* buf, int len, int flags)
 
 int ODSocket::Recv(char* buf, int len, int flags)
 {
-	return (recv(m_sock, buf, len, flags));
+	return (int)(recv(m_sock, buf, len, flags));
 }
 
 int ODSocket::Close()
